+++ sim/druid/forms.go
@@ -38,16 +38,6 @@ func (druid *Druid) GetCatWeapon() core.Weapon {
 	}
 }
 
-// Func (druid *Druid) GetBearWeapon() core.Weapon {
-// 	return core.Weapon{
-// 		BaseDamageMin:        109,
-// 		BaseDamageMax:        165,
-// 		SwingSpeed:           2.5,
-// 		NormalizedSwingSpeed: 2.5,
-// 		AttackPowerPerDPS:    core.DefaultAttackPowerPerDPS,
-// 	}
-// }
-
 // TODO: Class bonus stats for both cat and bear.
 func (druid *Druid) GetFormShiftStats() stats.Stats {
 	s := stats.Stats{
@@ -96,7 +86,7 @@ func (druid *Druid) registerCatFormSpell() {
 
 	var hotwDep *stats.StatDependency
 	if druid.Talents.HeartOfTheWild > 0 {
-		hotwDep = druid.NewDynamicMultiplyStat(stats.Strength, 1.0+0.04*float64(druid.Talents.HeartOfTheWild))
+		hotwDep = druid.NewDynamicMultiplyStat(stats.Strength, 1.0+0.03*float64(druid.Talents.HeartOfTheWild)) // DBC 17003-24894: 3%/rank
 	}
 
 	clawWeapon := druid.GetCatWeapon()
@@ -218,145 +208,6 @@ func (druid *Druid) registerCatFormSpell() {
 		},
 	})
 }
-
-// func (druid *Druid) registerBearFormSpell() {
-// 	actionID := core.ActionID{SpellID: 9634}
-// 	healthMetrics := druid.NewHealthMetrics(actionID)
-
-// 	statBonus := druid.GetFormShiftStats().Add(stats.Stats{
-// 		stats.AttackPower: 3 * float64(druid.Level),
-// 	})
-
-// 	stamDep := druid.NewDynamicMultiplyStat(stats.Stamina, 1.25)
-
-// 	var potpDep *stats.StatDependency
-// 	if druid.Talents.ProtectorOfThePack > 0 {
-// 		potpDep = druid.NewDynamicMultiplyStat(stats.AttackPower, 1.0+0.02*float64(druid.Talents.ProtectorOfThePack))
-// 	}
-
-// 	var hotwDep *stats.StatDependency
-// 	if druid.Talents.HeartOfTheWild > 0 {
-// 		hotwDep = druid.NewDynamicMultiplyStat(stats.Stamina, 1.0+0.02*float64(druid.Talents.HeartOfTheWild))
-// 	}
-
-// 	potpdtm := 1 - 0.04*float64(druid.Talents.ProtectorOfThePack)
-
-// 	clawWeapon := druid.GetBearWeapon()
-// 	predBonus := stats.Stats{}
-
-// 	druid.BearFormAura = druid.RegisterAura(core.Aura{
-// 		Label:      "Bear Form",
-// 		ActionID:   actionID,
-// 		Duration:   core.NeverExpires,
-// 		BuildPhase: core.Ternary(druid.StartingForm.Matches(Bear), core.CharacterBuildPhaseBase, core.CharacterBuildPhaseNone),
-// 		OnGain: func(aura *core.Aura, sim *core.Simulation) {
-// 			if !druid.Env.MeasuringStats && druid.form != Humanoid {
-// 				druid.CancelShapeshift(sim)
-// 			}
-// 			druid.form = Bear
-// 			druid.SetCurrentPowerBar(core.RageBar)
-
-// 			druid.AutoAttacks.SetMH(clawWeapon)
-
-// 			druid.PseudoStats.ThreatMultiplier *= 2.1021
-// 			druid.PseudoStats.SchoolDamageDealtMultiplier[stats.SchoolIndexPhysical] *= 1.0 + 0.02*float64(druid.Talents.MasterShapeshifter)
-// 			druid.PseudoStats.DamageTakenMultiplier *= potpdtm
-//			Switch to using AddStat as PseudoStat is being removed
-// 			druid.PseudoStats.BaseDodge += 0.02 * float64(druid.Talents.FeralSwiftness+druid.Talents.NaturalReaction)
-
-// 			predBonus = druid.GetDynamicPredStrikeStats()
-// 			druid.AddStatsDynamic(sim, predBonus)
-// 			druid.AddStatsDynamic(sim, statBonus)
-// 			druid.ApplyDynamicEquipScaling(sim, stats.Armor, druid.BearArmorMultiplier())
-// 			if potpDep != nil {
-// 				druid.EnableDynamicStatDep(sim, potpDep)
-// 			}
-
-// 			// Preserve fraction of max health when shifting
-// 			healthFrac := druid.CurrentHealth() / druid.MaxHealth()
-// 			druid.EnableDynamicStatDep(sim, stamDep)
-// 			if hotwDep != nil {
-// 				druid.EnableDynamicStatDep(sim, hotwDep)
-// 			}
-// 			druid.GainHealth(sim, healthFrac*druid.MaxHealth()-druid.CurrentHealth(), healthMetrics)
-
-// 			if !druid.Env.MeasuringStats {
-// 				druid.AutoAttacks.SetReplaceMHSwing(druid.ReplaceBearMHFunc)
-// 				druid.AutoAttacks.EnableAutoSwing(sim)
-
-// 				druid.manageCooldownsEnabled()
-// 				druid.UpdateManaRegenRates()
-// 			}
-// 		},
-// 		OnExpire: func(aura *core.Aura, sim *core.Simulation) {
-// 			druid.form = Humanoid
-// 			druid.AutoAttacks.SetMH(druid.WeaponFromMainHand())
-
-// 			druid.PseudoStats.ThreatMultiplier /= 2.1021
-// 			druid.PseudoStats.SchoolDamageDealtMultiplier[stats.SchoolIndexPhysical] /= 1.0 + 0.02*float64(druid.Talents.MasterShapeshifter)
-// 			druid.PseudoStats.DamageTakenMultiplier /= potpdtm
-//			Switch to using AddStat as PseudoStat is being removed
-// 			druid.PseudoStats.BaseDodge -= 0.02 * float64(druid.Talents.FeralSwiftness+druid.Talents.NaturalReaction)
-
-// 			druid.AddStatsDynamic(sim, predBonus.Invert())
-// 			druid.AddStatsDynamic(sim, statBonus.Invert())
-// 			druid.RemoveDynamicEquipScaling(sim, stats.Armor, druid.BearArmorMultiplier())
-// 			if potpDep != nil {
-// 				druid.DisableDynamicStatDep(sim, potpDep)
-// 			}
-
-// 			healthFrac := druid.CurrentHealth() / druid.MaxHealth()
-// 			druid.DisableDynamicStatDep(sim, stamDep)
-// 			if hotwDep != nil {
-// 				druid.DisableDynamicStatDep(sim, hotwDep)
-// 			}
-// 			druid.RemoveHealth(sim, druid.CurrentHealth()-healthFrac*druid.MaxHealth())
-
-// 			if !druid.Env.MeasuringStats {
-// 				druid.AutoAttacks.SetReplaceMHSwing(nil)
-// 				druid.AutoAttacks.EnableAutoSwing(sim)
-
-// 				druid.manageCooldownsEnabled()
-// 				druid.UpdateManaRegenRates()
-// 				druid.EnrageAura.Deactivate(sim)
-// 				druid.MaulQueueAura.Deactivate(sim)
-// 			}
-// 		},
-// 	})
-
-// 	rageMetrics := druid.NewRageMetrics(actionID)
-
-// 	furorProcChance := []float64{0, 0.2, 0.4, 0.6, 0.8, 1}[druid.Talents.Furor]
-
-// 	druid.BearForm = druid.RegisterSpell(Any, core.SpellConfig{
-// 		ActionID: actionID,
-// 		Flags:    core.SpellFlagNoOnCastComplete | core.SpellFlagAPL,
-
-// 		ManaCost: core.ManaCostOptions{
-// 			BaseCost:   0.55,
-// 			Multiplier: (1 - 0.2*float64(druid.Talents.KingOfTheJungle)) * (1 - 0.1*float64(druid.Talents.NaturalShapeshifter)),
-// 		},
-// 		Cast: core.CastConfig{
-// 			DefaultCast: core.Cast{
-// 				GCD: core.GCDDefault,
-// 			},
-// 			IgnoreHaste: true,
-// 		},
-
-// 		ApplyEffects: func(sim *core.Simulation, _ *core.Unit, spell *core.Spell) {
-// 			rageDelta := 0 - druid.CurrentRage()
-// 			if sim.Proc(furorProcChance, "Furor") {
-// 				rageDelta += 10
-// 			}
-// 			if rageDelta > 0 {
-// 				druid.AddRage(sim, rageDelta, rageMetrics)
-// 			} else if rageDelta < 0 {
-// 				druid.SpendRage(sim, -rageDelta, rageMetrics)
-// 			}
-// 			druid.BearFormAura.Activate(sim)
-// 		},
-// 	})
-// }
 
 func (druid *Druid) manageCooldownsEnabled() {
 	// Disable cooldowns not usable in form and/or delay others