+++ sim/druid/tank/tank.go
@@ -38,17 +38,18 @@ func NewFeralTankDruid(character *core.Character, options *proto.Player) *FeralT
 	}
 
 	bear.EnableRageBar(core.RageBarOptions{
-		StartingRage:   bear.Options.StartingRage,
-		RageMultiplier: 1,
+		StartingRage:          bear.Options.StartingRage,
+		DamageDealtMultiplier: 1,
+		DamageTakenMultiplier: 1,
 	})
 
+	bear.ReplaceBearMHFunc = bear.TryMaul
 	bear.EnableAutoAttacks(bear, core.AutoAttackOptions{
 		// Base paw weapon.
 		MainHand:       bear.GetBearWeapon(),
 		AutoSwingMelee: true,
 		ReplaceMHSwing: bear.TryMaul,
 	})
-	bear.ReplaceBearMHFunc = bear.TryMaul
 
 	healingModel := options.HealingModel
 	if healingModel != nil {
@@ -77,7 +78,7 @@ func (bear *FeralTankDruid) Initialize() {
 
 func (bear *FeralTankDruid) Reset(sim *core.Simulation) {
 	bear.Druid.Reset(sim)
-	bear.Druid.ClearForm(sim)
+	bear.Druid.CancelShapeshift(sim)
 	bear.BearFormAura.Activate(sim)
 	bear.Druid.PseudoStats.Stunned = false
 }