+++ ui/feral_tank_druid/inputs.ts
@@ -11,22 +11,3 @@ export const StartingRage = InputHelpers.makeSpecOptionsNumberInput<Spec.SpecFer
 	labelTooltip: 'Initial rage at the start of each iteration.',
 });
 
-export const FeralTankDruidRotationConfig = {
-	inputs: [
-		InputHelpers.makeRotationNumberInput<Spec.SpecFeralTankDruid>({
-			fieldName: 'maulRageThreshold',
-			label: 'Maul Rage Threshold',
-			labelTooltip: 'Queue Maul when Rage is above this value.',
-		}),
-		InputHelpers.makeRotationNumberInput<Spec.SpecFeralTankDruid>({
-			fieldName: 'lacerateTime',
-			label: 'Lacerate Refresh Leeway',
-			labelTooltip: 'Refresh Lacerate when remaining duration is less than this value (in seconds).',
-		}),
-		InputHelpers.makeRotationBooleanInput<Spec.SpecFeralTankDruid>({
-			fieldName: 'maintainDemoralizingRoar',
-			label: 'Maintain Demo Roar',
-			labelTooltip: 'Keep Demoralizing Roar active on the primary target. If a stronger debuff is active, will not cast.',
-		}),
-	],
-};